merge: Tighten error bounds, explain reasoning
After thinking about it overnight, I realized that we should be able to
guarantee a tighter error bound when merging. This should substantially
improve the error bounds for long chains of merges.

Note that I _still_ haven't written a rigorous proof or anything. This
is based on my understanding of the algorithm, which isn't guaranteed
to be 100% complete. However, this was tested with 100,000 random test
cases:

    PROPTEST_CASES=100000 cargo test test_merge_lemma3_invariants

...with no failures. Broken versions have typically failed within 100
iterations, so 100000 offers us some decent confidence.

I have also added comments with my understanding of why this should
work. But honestly, the proptests are carrying some real weight in
convincing me that we're upholding our invariants and that I haven't
fallen prey to insufficient coffee when reasoning through it.